Update dmg-license.py from pyhacker and include license info for it
